requirements.txt
torch>=2.4.0
transformers>=4.44.0
peft>=0.12.0
datasets>=2.21.0
accelerate>=0.33.0
bitsandbytes>=0.43.3
# llama-cpp-python>=0.2.90  <-- Handled separately in setup.ps1
sentencepiece>=0.2.0
safetensors>=0.4.4
chromadb>=0.5.5
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
gradio>=4.42.0
huggingface_hub>=0.24.6
protobuf>=5.27.3
scipy>=1.14.1
tiktoken>=0.7.0
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
pypdf>=4.0.0

# Visualization & Analysis Tools
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0
scikit-learn>=1.3.0
wandb>=0.15.0
tensorboard>=2.14.0

# Data Processing
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
nltk>=3.8.0
rouge-score>=0.1.2
sacrebleu>=2.3.0

# Development & Formatting
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
streamlit>=1.28.0
